fix: allow using construct in other construct libraries
